moves sign out to nav bar
