+++ src/Components/Navbar.js
@@ -1,6 +1,12 @@
 import { Link } from "react-router-dom";
+import { supabase } from "../supabaseClient";
 
 export default function Navbar({ session }) {
+  const handleSignOut = async () => {
+    await supabase.auth.signOut();
+    window.location.href = "http://localhost:3000/login";
+  };
+
   return (
     <div>
       {!session ? (
@@ -10,6 +16,13 @@ export default function Navbar({ session }) {
       ) : (
         <div>
           <Link to="/account">my account</Link>
+          <button
+            type="button"
+            className="button block"
+            onClick={handleSignOut}
+          >
+            Sign Out
+          </button>
         </div>
       )}
     </div>
